Replace freezable's usage of ActiveSupport::Concern with plain ruby
Because we were freezing parent classes (like Object), included hooks
were not being invoked as expected.

It also extracts refrigerator to deal with class-freezing
