+++ lib/console1984/shield.rb
@@ -19,7 +19,7 @@ class Console1984::Shield
   # that aren't mean to be modified once the console is running.
   def install
     extend_protected_systems
-    freeze_all
+    refrigerator.freeze_all
   end
 
   private
@@ -65,16 +65,8 @@ class Console1984::Shield
       end
     end
 
-    def freeze_all
-      eager_load_all_classes
-      Console1984.config.freeze unless Console1984.config.test_mode
-      Console1984::Freezeable.freeze_all
-      Parser::CurrentRuby.freeze
-    end
-
-    def eager_load_all_classes
-      Rails.application.eager_load! unless Rails.application.config.eager_load
-      Console1984.class_loader.eager_load
+    def refrigerator
+      @refrigerator ||= Console1984::Refrigerator.new
     end
 
     module SSLSocketRemoteAddress
