+++ test/tampering_cases/audit_models/tamper_parser.rb
@@ -1,0 +1,4 @@
+class Parser::CurrentRuby
+  def omg
+  end
+end
